Fixed Game Update Rate
The frog was limited to only making one move every fraction of a second
as without the limitation it moved too fast. However, because it was
time based, when the population is 1000 frogs large to the point the
game is laggy, the frogs can move extremely fast relative to the cars,
almost as if the cars are not moving. To fix this the frogs being able
to move is related to gameTicks instead of timeChanged