Genetic Algorithm Incorporated
Uses the same mutation/Crossover methods as the Asteroids example.
Population set up is essentially the same. Frogs auto die after so much
time to ensure they are not caught in a left/right loop. Can develop
some interesting behavior, but does not always. The fitness is extremely
simple and it as long as the crossover/mutation methods need changed to
find optimal AI.